Split terminal raw mode and signal handling into platform-specific files
TCGETS/TCSETS are Linux-only ioctl constants; macOS uses TIOCGETA/TIOCSETA.
SIGWINCH is unavailable on Windows. Extract enableOutputProcessing into
raw_linux.go, raw_darwin.go, raw_windows.go and sigwinchLoop into
signal_unix.go, signal_windows.go to allow cross-compilation.
