+++ service/terminal/raw_windows.go
@@ -1,0 +1,7 @@
+//go:build windows
+
+package terminal
+
+// Windows console does not use POSIX termios; output processing
+// is handled by the console subsystem.
+func enableOutputProcessing(fd int) {}
