+++ service/terminal/signal_windows.go
@@ -1,0 +1,10 @@
+//go:build windows
+
+package terminal
+
+import "context"
+
+func (r *InputReader) sigwinchLoop(ctx context.Context) {
+	defer r.wg.Done()
+	<-ctx.Done()
+}
